testing_api/requirements.txt
# FastAPI Testing API Requirements

# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6  # For file uploads

# Pydantic for data validation
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# MongoDB
pymongo>=4.6.0
